added package summary
