+++ .gitignore
@@ -3,3 +3,5 @@
 
 # A little hack I use to see the output of a proc macro, even when it fails to compile
 proc-macro-out.rs
+
+export-example.json
