+++ Cargo.toml
@@ -38,6 +38,7 @@ serde_json = { version = "1.0", features = ["preserve_order"] }
 sfsu-derive = { path = "./sfsu-derive" }
 strum = { version = "0.25", features = ["derive"] }
 thiserror = "1.0"
+url = { version = "2.5", features = ["serde"] }
 
 [dev-dependencies]
 criterion = { version = "0.5", features = ["html_reports"] }
