+++ src/lib.rs
@@ -8,6 +8,7 @@ use rayon::prelude::*;
 pub mod buckets;
 pub mod config;
 pub mod packages;
+pub mod summary;
 
 mod opt;
 /// Currently this is mostly an internal api
