+++ src/summary.rs
@@ -1,0 +1,4 @@
+//! Summary info structs
+
+pub mod bucket;
+pub mod package;
